Initial commit
So far we have the basic skeleton for the Plasma-based wizard and
and couple pages to show how it works and looks like.

The ColorThemePage does not work properly yet and there's no integration
with PackageKit (I have to learn how to use the API first :-)).

We also need a class to collect user's configuration so that we can put
up a summary at the end and we have a list of changes to apply.

TODO:
 - polish UI
 - pre-load pages (switching tabs is awfuly slow atm)
 - look into the 'QPainter' warnings in console
 - test look&feel properly with default Plasma configuration
 - add more pages
 - create a proper TODO file

Note: to make this work properly, you have to copy content of 'data'
directory to /usr/share/kde4/apps/fedora-plasma-first-login/ directory.
I don't have time to write install() section now :-)