Reimplemented Regions page in QML
I haven't tested it properly, because for some reason I'm getting empty
list from KGlobal::locale()->allCountriesList(), but it's pretty much
a copy of Langages page, so it should work.

Since they share so much code, it might be worth factoring out all the
common code, which is pretty much everything except for the source
model and couple labels.